src/requirements.txt
Django==3.1.*
psycopg2-binary
djangorestframework==3.12.*
django-filter==2.4.*
pyjwt[crypto]==2.1.*
pycryptodome==3.10.*

# Dev
flake8==3.8.*
isort[colors]
django_extensions
uritemplate
pyyaml
